Add weak map to hold static methods
Static methods are now stored in a weak map for compose
factories so that when static methods are being 'passed'
from one factory to another only the explicitly defined
static methods are copied.

Added test cases to verify static behavior for compose
factories and added some words about static to the
documentation.